mlflow-scripts/requirements.txt
# requirements.txt
lightning>=2.0.0
mlflow>=2.0.0
torch>=1.13.0
torchvision>=0.14.0
kagglehub>=0.2.1
torchmetrics>=0.11.0
matplotlib>=3.5.0
pillow>=9.0.0
numpy>=1.21.0
scikit-image>=0.19.0
# Branching of transformers needed for DepthPro
transformers @ git+https://github.com/geetu040/transformers.git@depth-pro-projects
